fix(worker): register a session as a tool-server user before its start can time out
start_session_tool_servers spawned ServerReconciler::session_started into a
background task and timed out the join, but session_started registered the
session as a user (in claim_or_wait) and started the actual process in the
same call. On timeout that task kept running by design; if
end_session_tool_servers ran first (ack failure, or a fast session ending
before the delayed task got to it), session_ended found nothing to release,
and the registration that landed afterward pinned the server as a user-less
'running' entry for the rest of the worker's lifetime.

Split ServerReconciler::session_started into claim_users (registration,
always awaited directly and to completion) and start_claimed (the slow
part: spawning the process and waiting for its registration). Only
start_claimed is what start_session_tool_servers now times out and
detaches into the background; claim_users always finishes before
start_session_tool_servers returns, so end_session_tool_servers can never
run before its matching registration exists.

Also split tests/tool_reconciler.rs into tool_reconciler.rs (plain
refcount/linger rules), tool_reconciler_race.rs (teardown/start races and
concurrency, including the new claim_users regression), and
tool_reconciler_e2e.rs (real worker, real broker) — the new race test
pushed the single file over CodeScene's cohesion threshold.
